src/requirements.txt
discord-py-interactions==5.11.0
PyYAML==6.0.2
psycopg2==2.9.10